+++ core/src/main/java/cards/utils/AttackAble.java
@@ -9,7 +9,7 @@ public interface AttackAble {
      *
      * @param damage the damage
      */
-    void reduceHealthBy(int damage);
+    void reduceHealthBy(double damage);
 
     /**
      * get the type of the attacker itself.
